fix for constant length sequences
This fixes error which occurs when parsing Eurex messages with mFAST.

Eurex schema uses constant length sequences in several places, e.g. in
QuoteRequest message:
```
  <sequence name="SequenceGroup">
      <length name="NoElements" id="111">
        <constant value="1"/>
      </length>
      <int64 name="Id" id="112">
        <copy/>
      </int64>
    </sequence>
```

Unfortunately mFAST fails to properly handle that schema.

The problem boils down to code decoding sequence instruction:
```
    template <typename T>
    inline void fast_decoder_base::decode_field(const T &ext_ref,
                                                sequence_type_tag) {
      value_storage storage;

      auto length = ext_ref.set_length(storage);
      this->visit(length);
```
Here value_storage is not initialized with initial (constant) value.
Later, in
```
    template <typename T, typename TypeCategory>
    void fast_decoder_base::decode_field(const T &ext_ref,
                                         constant_operator_tag,
                                         TypeCategory) {
```
the value for constant field is not copied (and that's what was changed
in original patch) and sequence length is considered to be 0 leaving all
of sequence payload bytes unparsed.